Algoritmos/SERVICIOS-AURALIS/auralis-subscriber final/requirements.txt
paho-mqtt>=1.6.1
PyMySQL>=1.1.0
python-dotenv>=1.0.1
